Indentation and edebug support for threading macros
Given an elisp expression of the form:

    (->> (+ 1 1)
         (+ 2)
         (* 3)
         (format "%s"))

This tells edebug that the first argument is an expression, so users can
step through it. This is an improvement over the current situation,
where users cannot step through any part of a threaded expression.

I've also changed the indentation, so threading macros are indented
similarly to other scoping macros:

    (->> (+ 1 1)
      (+ 2)
      (* 3)
      (format "%s"))

Note that this indentation and debugging behaviour matches thread-first
in Emacs 25:
http://git.savannah.gnu.org/cgit/emacs.git/tree/lisp/emacs-lisp/subr-x.el?id=emacs-25.0.95#n49